+++ cron/cron-spec/index.ts
@@ -4,7 +4,7 @@ import {
   type ExtractRequestBody,
   type ExtractResponseBody,
   castJson,
-} from "@saflib/openapi-specs";
+} from "@saflib/openapi";
 
 // Export the JSON spec for middleware
 export const jsonSpec = castJson(json);
