+++ openapi/helpers.ts
